@@ -1,15 +1,13 @@
 package step.learning.oop;
 
-//import java.text.SimpleDateFormat;
-//import java.time.format.DateTimeFormatter; // Import the DateTimeFormatter class
+import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 
-public class Newspaper extends Literature implements Periodic{
-    //private Date date;
+public class Hologram extends Literature implements Expo{
     private LocalDate date;
     //private static final SimpleDateFormat dateFormat = new SimpleDateFormat("dd.MM.yyyy");
 
-    public Newspaper(String title, LocalDate date) {
+    public Hologram(String title, LocalDate date) {
         super.setTitle(title);
         this.setDate(date);
     }
@@ -24,9 +22,6 @@ public class Newspaper extends Literature implements Periodic{
 
     @Override
     public String getCard() {
-        return String.format("Newspaper: %s, date: %s.", this.getTitle(), this.getDate());
-    }
-    public String getPeriod() {
-        return "daily";
+        return String.format("Hologram: %s, date: %s.", this.getTitle(), this.getDate());
     }
 }
